[Feature][list] Refactoring packages for MVC
