+++ app/src/androidTest/java/com/example/TopMovieList/ExampleInstrumentedTest.java
@@ -3,7 +3,7 @@
  * All Rights Reserved
  */
 
-package com.example.mobileproject;
+package com.example.TopMovieList;
 
 import android.content.Context;
