+++ app/src/main/java/com/example/TopMovieList/Constants.java
@@ -1,0 +1,15 @@
+/*
+ * Copyright (c) 2020. Antoine Mairet
+ * All Rights Reserved
+ */
+
+package com.example.TopMovieList;
+
+//Class used to store constant value
+public class Constants {
+    public static String KEY_MOVIE_LIST = "jsonListMovie";
+    public static final String BASE_URL = "https://imdb-api.com/";
+    public static String KEY_MOVIE_WATCHLIST = "jsonWatchListMovie";
+    public static String KEY_MOVIE_FROM_MAIN_TO_DETAILS = "jsonMovieFromMainToDetails";
+    public static String KEY_MOVIE_FROM_DETAILS_TO_WATCHLIST = "jsonMovieFromDetailsToWatchList";
+}
