+++ app/src/main/java/com/example/TopMovieList/presentation/model/RestMoviesResponse.java
@@ -3,7 +3,7 @@
  * All Rights Reserved
  */
 
-package com.example.mobileproject;
+package com.example.TopMovieList.presentation.model;
 
 import java.util.ArrayList;
