+++ app/src/main/java/com/example/TopMovieList/presentation/view/WatchListActivity.java
@@ -3,7 +3,7 @@
  * All Rights Reserved
  */
 
-package com.example.mobileproject;
+package com.example.TopMovieList.presentation.view;
 
 import android.content.Context;
 import android.content.SharedPreferences;
@@ -14,6 +14,10 @@ import android.widget.Toast;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.example.TopMovieList.Constants;
+import com.example.TopMovieList.R;
+import com.example.TopMovieList.presentation.model.Movie;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.reflect.TypeToken;
