--- app/src/main/java/com/example/mobileproject/Constants.java
@@ -1,15 +1,0 @@
-/*
- * Copyright (c) 2020. Antoine Mairet
- * All Rights Reserved
- */
-
-package com.example.mobileproject;
-
-//Class used to store constant value
-public class Constants {
-    static String KEY_MOVIE_LIST = "jsonListMovie";
-    static final String BASE_URL = "https://imdb-api.com/";
-    static String KEY_MOVIE_WATCHLIST = "jsonWatchListMovie";
-    static String KEY_MOVIE_FROM_MAIN_TO_DETAILS = "jsonMovieFromMainToDetails";
-    static String KEY_MOVIE_FROM_DETAILS_TO_WATCHLIST = "jsonMovieFromDetailsToWatchList";
-}
