+++ app/src/main/res/layout/activity_main.xml
@@ -9,7 +9,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainListActivity"
+    tools:context=".presentation.view.MainListActivity"
     android:background="@drawable/gradient_background">
 
     <androidx.recyclerview.widget.RecyclerView
